requirements.txt
Pillow
rembg
pynput
pystray
pywin32